Remove `build-image-index` task
Task bundles are architecture-neutral containers with data and so
`build-image-index` is currently disabled and I'm confident we
will not need to enable it in the foreseeable future.

If some Tekton task needs to use some native binaries, these are
provided through step's `image:` attribute.

Therefore `build-image-index` is simply redundant at this point.