Tell Renovate to schedule as frequently as possible
This repo requires special setup so at this point we need to see
the effects of changes as soon as possible to not delay the
iteration.
My suggestion is to merge with `at any time` schedule for now
because only config in the main branch will be effective.

Later, as part of the same ROX-27352 ticket I will change the
schedule to our normal "during night hours".