Create a task for figuring Collector+Scanner image tags
I tried having this logic as part of the regular
`determine-image-tag` task but it turned to be oddly-looking when
I started adding tagging support and so here I just move it to its
own place to live its own life.

Note tha this task doesn't call `fail-build-if-git-is-dirty.sh`
script because the outputs aren't affected by whether git changes
are committed.