Remove unexpected backticks
```
Tag from Tekton event: ''
Tag reported by Makefile: 'misha-test-8'
Tags seen by git: 'misha-test-8'
Long git describe output: 'misha-test-8-0-g8d18e3b812'
This is not a tag push event but Makefile reports literally the git tag 'misha-test-8'.
This happens when a build was triggered not by a tag push event but the commit is tagged and when the Makefile doesn't use '--long' with 'git describe'.
/tekton/scripts/script-1-7v9vw: line 73: misha-test-8: command not found
We should use a different image tag for this build in order to not mix results with a build that was triggered by the tag push event and which will indeed use  as the tag for images built there.
Using misha-test-8-0-g8d18e3b812 for the tag.
misha-test-8-0-g8d18e3b812-fast
```